システムテストを受講生用とメンター用で分けた
